Background formulation and causal bottlenecks.
DO_BACKGROUND is a mode where the first #BACKGROUND_LEN tokens in each
sample are background tokens; During training we don't enforce a loss on
the targets in the background. During sampling we always have the background.

DO_CAUSAL_BOTTLENECK is a GPT modification where we append #STYLE_VECS_LEN
tokens to the background, and add edit the causal mask such that generated
tokens can see the style vectors but not the background itself. Hence the
information from the background can only be passed through the states produced
in the style vector timesteps.

Requires slight mods to RPE:
  Compute shifts to rotary embeddings to simulate the embeddings
  we'd see at different starting timesteps.